+++ crates/holochain_sqlite/src/sql/cell/fetch_hashes/any/fetch_op_hashes_continuous.sql
@@ -2,11 +2,11 @@
 -- i.e. when :storage_start_loc <= :storage_end_loc
 SELECT
   hash,
-  authored_timestamp_ms
+  authored_timestamp
 FROM
   DHtOp
 WHERE
-  DhtOp.authored_timestamp_ms >= :from
-  AND DhtOp.authored_timestamp_ms < :to
+  DhtOp.authored_timestamp >= :from
+  AND DhtOp.authored_timestamp < :to
   AND storage_center_loc >= :storage_start_loc
   AND storage_center_loc <= :storage_end_loc