+++ crates/holochain_sqlite/src/sql/cell/fetch_hashes/integrated/fetch_op_hashes_wrapped_v2.sql
@@ -8,8 +8,8 @@ FROM
   DHtOp
 WHERE
   DhtOp.when_integrated IS NOT NULL
-  AND DhtOp.authored_timestamp_ms >= :from
-  AND DhtOp.authored_timestamp_ms < :to
+  AND DhtOp.authored_timestamp >= :from
+  AND DhtOp.authored_timestamp < :to
   AND storage_center_loc < :storage_end_loc
 UNION
 ALL
@@ -19,6 +19,6 @@ FROM
   DHtOp
 WHERE
   DhtOp.when_integrated IS NOT NULL
-  AND DhtOp.authored_timestamp_ms >= :from
-  AND DhtOp.authored_timestamp_ms < :to
+  AND DhtOp.authored_timestamp >= :from
+  AND DhtOp.authored_timestamp < :to
   AND storage_center_loc > :storage_start_loc