+++ crates/kitsune_p2p/kitsune_p2p/src/types/event.rs
@@ -1,6 +1,7 @@
 //! Definitions for events emited from the KitsuneP2p actor.
 
 use crate::types::agent_store::AgentInfoSigned;
+use kitsune_p2p_timestamp::Timestamp;
 use kitsune_p2p_types::dht_arc::DhtArcSet;
 use std::{collections::HashSet, sync::Arc, time::SystemTime};
 
@@ -14,7 +15,7 @@ pub struct QueryOpHashesEvt {
     /// The agents from which to fetch, along with a DhtArcSet to filter by.
     pub agents: Vec<(KAgent, DhtArcSet)>,
     /// The time window to search within.
-    pub window_ms: TimeWindowMs,
+    pub window: TimeWindow,
     /// Maximum number of ops to return.
     pub max_ops: usize,
     /// Include ops that are still in limbo (not yet validated or integrated).
@@ -155,15 +156,12 @@ pub enum MetricQueryAnswer {
     Oldest(Option<KAgent>),
 }
 
-/// A UNIX timestamp, measured in milliseconds
-pub type TimestampMs = u64;
-
 /// A range of timestamps, measured in milliseconds
-pub type TimeWindowMs = std::ops::Range<TimestampMs>;
+pub type TimeWindow = std::ops::Range<Timestamp>;
 
 /// A time window which covers all of recordable time
-pub fn full_time_window() -> TimeWindowMs {
-    TimestampMs::MIN..TimestampMs::MAX
+pub fn full_time_range() -> TimeWindow {
+    Timestamp::MIN..Timestamp::MAX
 }
 type KSpace = Arc<super::KitsuneSpace>;
 type KAgent = Arc<super::KitsuneAgent>;
@@ -211,7 +209,7 @@ ghost_actor::ghost_chan! {
         /// Gather a list of op-hashes from our implementor that meet criteria.
         /// Get the oldest and newest times for ops within a time window and max number of ops.
         // maackle: do we really need to *individually* wrap all these op hashes in Arcs?
-        fn query_op_hashes(input: QueryOpHashesEvt) -> Option<(Vec<KOpHash>, TimeWindowMs)>;
+        fn query_op_hashes(input: QueryOpHashesEvt) -> Option<(Vec<KOpHash>, TimeWindow)>;
 
         /// Gather all op-hash data for a list of op-hashes from our implementor.
         fn fetch_op_data(input: FetchOpDataEvt) -> Vec<(KOpHash, Vec<u8>)>;